added some guard to status delete

+++ controllers/appController.js
@@ -1,3 +1,4 @@
+const Status = require('../models/Status');
 const {createStatus, deleteStatus} = require('../services/app');
 
 const router = require('express').Router();
@@ -14,12 +15,20 @@ router.post('/status/create', async (req, res) => {
 
 router.delete('/status/:id/delete', async (req, res) => {
   try {
-    await deleteStatus(req.params.id);
-    
+    const status = await Status.findById(req.params.id);
+
+    if (status._ownerId == req.user._id) {
+      await deleteStatus(req.params.id);
+    } else {
+      throw new Error('You cannot delete this status.');
+    }
+
     res.status(200).json('Deleted');
   } catch (err) {
     res.status(400).json({error: err.message});
   }
 });
+
+
 
 module.exports = router;
